Harvester_T/requirements.txt
couchdb==1.2
nltk==3.7
pandas==1.4.2
python-dotenv==0.20.0
snscrape==0.4.3.20220106
tweepy==4.8.0
